requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
efficientnet_pytorch
facenet-pytorch
opencv-python
librosa
soundfile
numpy
pandas
scikit-learn
matplotlib
seaborn
fastapi
uvicorn[standard]
python-multipart
Pillow
tqdm